Separate present feature as StateAndBalanceFeature
