+++ nekoyume/Assets/Plugins/StateViewer/Tests/EditMode/StateTreeViewItemModelTest.cs
@@ -1,13 +1,15 @@
 using Bencodex.Types;
 using NUnit.Framework;
 using StateViewer.Editor;
+using StateViewer.Editor.Features;
 using StateViewer.Runtime;
 
 namespace StateViewer.Tests.EditMode
 {
     public class StateTreeViewItemModelTest
     {
-        [Test, TestCaseSource(typeof(StateViewerWindow), nameof(StateViewerWindow.TestValues))]
+        [Test,
+         TestCaseSource(typeof(StateViewerWindow), nameof(StateAndBalanceFeature.TestValues))]
         public void Serialize(IValue value)
         {
             // Arrange
